+++ playroom/snippets.tsx
@@ -117,8 +117,8 @@ const formSnippets: Array<Snippet> = [
             '  </Stack>\n' +
             '</RadioGroup>',
     ],
-    ['OtpField', '<OtpField name="otp" aria-label="OTP" />'],
-    ['OtpField (hideCode)', '<OtpField hideCode name="pin" aria-label="PIN" />'],
+    ['PinField', '<PinField name="otp" aria-label="OTP" />'],
+    ['PinField (hideCode)', '<PinField hideCode name="pin" aria-label="PIN" />'],
     [
         'Form',
         `<Form
