+++ src/__screenshot_tests__/__image_snapshots__/form-fields-screenshot-test-tsx-pin-field-2-snap.png
