+++ src/__screenshot_tests__/__image_snapshots__/form-fields-screenshot-test-tsx-pin-field-hide-code-1-snap.png
