+++ src/theme.tsx
@@ -41,7 +41,7 @@ const TEXTS_ES = {
     playIconButtonLabel: 'Reproducir',
     pauseIconButtonLabel: 'Pausar',
     sheetConfirmButton: 'Continuar',
-    otpFieldInputLabel: 'Dígito 1$s de 2$s',
+    pinFieldInputLabel: 'Dígito 1$s de 2$s',
 };
 
 const TEXTS_EN: ThemeTexts = {
@@ -78,7 +78,7 @@ const TEXTS_EN: ThemeTexts = {
     playIconButtonLabel: 'Play',
     pauseIconButtonLabel: 'Pause',
     sheetConfirmButton: 'Continue',
-    otpFieldInputLabel: 'Digit 1$s of 2$s',
+    pinFieldInputLabel: 'Digit 1$s of 2$s',
 };
 
 const TEXTS_DE: ThemeTexts = {
@@ -115,7 +115,7 @@ const TEXTS_DE: ThemeTexts = {
     playIconButtonLabel: 'Abspielen',
     pauseIconButtonLabel: 'Pausieren',
     sheetConfirmButton: 'Fortfahren',
-    otpFieldInputLabel: 'Ziffer 1$s von 2$s',
+    pinFieldInputLabel: 'Ziffer 1$s von 2$s',
 };
 
 const TEXTS_PT: ThemeTexts = {
@@ -152,7 +152,7 @@ const TEXTS_PT: ThemeTexts = {
     playIconButtonLabel: 'Reproduzir',
     pauseIconButtonLabel: 'Pausar',
     sheetConfirmButton: 'Continuar',
-    otpFieldInputLabel: 'Dígito 1$s de 2$s',
+    pinFieldInputLabel: 'Dígito 1$s de 2$s',
 };
 
 export const getTexts = (locale: Locale): typeof TEXTS_ES => {
